moved Navigation to common
